Feature: widen session_id to u64 and bump the wire protocol to v3.
- Widen frame_header.session_id u8->u64 end-to-end (codec offsets, the epoch_source well, peer_session members/accessors, and the forwarder session_id params); header_size 21->28.
- Pin the epoch mint width with a static_assert so a narrowing can never reintroduce the 255-reconnect wrap collision.
- Bump k_protocol_version 2->3 exactly once; the handshake payload is unchanged, and a skewed peer is rejected by the exact-match version gate.
- Allocate and reserve the gid-present flag bit for the source-identity carriage; a flag-clear v3 frame is byte-identical to a v3-no-flag frame.
- Prove the wrap is gone with a direct epoch_source unit test and a staleness-gate case distinguishing epochs that alias under u8.
- Update the wire-codec, handshake, and reconnect test pins to the v3 layout in lockstep.

+++ lib/plexus/include/plexus/io/message_forwarder.h
@@ -142,7 +142,7 @@ public:
     // 0 — passed per send (NOT a forwarder-wide member) because a node-shared
     // forwarder fans to many peers, each with its own epoch.
     void publish(std::string_view fqn, std::span<const std::byte> payload,
-                 std::uint8_t session_id = 0)
+                 std::uint64_t session_id = 0)
     {
         auto hash = wire::fqn_topic_hash(fqn);
         const auto *subs = m_registry.subscribers_for(hash);
